Até a pagina 64

+++ TestC/TestC/Models/Client.cs
@@ -7,10 +7,10 @@ namespace TestC.Models
 {
     public class Client
     {
-        public int ClientId { get; set; }
+        public long ClientId { get; set; }
         public string Name { get; set; }
         public int Cep { get; set; }
-        public int Cpf { get; set; }
+        public long Cpf { get; set; }
         public string Adress { get; set; }
         public string Email { get; set; }
         public int Fone { get; set; }
